# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sample API for get books info",
    "description": "This is a sample server for a book store.",
    "termsOfService": "http://example.com/terms/",
    "contact": {
      "name": "API Support",
      "email": "hmservicostech@outlook.com.br"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API Test"
    },
    {
      "url": "http://localhost:3001",
      "description": "API PRD"
    }
  ],
  "paths": {
    "/books": {
      "post": {
        "summary": "create books",
        "description": "this route will create a book",
        "tags": ["Books"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              },
              "examples": {
                "book": {
                  "value": {
                    "title": "Clean Code",
                    "author": "Robert Cecil Martin",
                    "publisher": "Prentice Hall PTR",
                    "numberPages": 431
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              }
            }
          },
          "400": {
            "description": "error"
          }
        }
      },
      "get": {
        "description": "get all books",
        "summary": "get all book",
        "tags": ["Books"],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "description": "get book by id",
        "summary": "get book by id",
        "tags": ["Books"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "book id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "change book by id",
        "summary": "change book by id",
        "tags": ["Books"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "book id",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Book"
              },
              "examples": {
                "book": {
                  "value": {
                    "title": "Harry Potter and the Philosopher's STONE"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Book"
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "delete book by id",
        "summary": "delete book by id",
        "tags": ["Books"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "book id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/book/findByTitle": {
      "get": {
        "operationId": "getPetsById",
        "description": "get book by title",
        "summary": "get book by title",
        "tags": ["Books"],
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "description": "book title",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "author": {
            "type": "string"
          },
          "publisher": {
            "type": "string"
          },
          "numberPages": {
            "type": "number"
          }
        }
      }
    }
  }
}